+++ internal/meta/gps.go
@@ -107,13 +107,13 @@ func ParseFloat(s string) float64 {
 }
 
 // NormalizeGPS normalizes the longitude and latitude of the GPS position to a generally valid range.
-func NormalizeGPS(lat, lng float64) (float32, float32) {
+func NormalizeGPS(lat, lng float64) (float64, float64) {
 	if lat < LatMax || lat > LatMax || lng < LngMax || lng > LngMax {
 		// Clip the latitude. Normalise the longitude.
 		lat, lng = clipLat(lat), normalizeLng(lng)
 	}
 
-	return float32(lat), float32(lng)
+	return lat, lng
 }
 
 func clipLat(lat float64) float64 {
